infotext_init: drop (int) cast on u8->u16 arg (93.8->100%)
objSetHintTextIdx takes a u16 param; the (int)(u8)def[0x19] spelling forced a
clrlwi r4,r0,16 at the call site. Dropping the (int) cast lets the lbz feed the
u16 arg directly (no mask), matching target.

(cherry picked from commit e5b0db67a30446410c6312705829b3e799aa3f55)

+++ src/main/dll/dll_0121_infotext.c
@@ -16,7 +16,7 @@ void infotext_init(int obj, s8* def)
     v = (u32)((GameObject*)obj)->objectFlags | 0x6000;
     ((GameObject*)obj)->objectFlags = v;
     ((GameObject*)obj)->anim.rotX = (s16)((s32)(u8)def[0x18] << 8);
-    objSetHintTextIdx(obj, (int)(u8)def[0x19]);
+    objSetHintTextIdx(obj, (u8)def[0x19]);
 }
 
 void cctestinfot_init(int obj, s8* def);
